german grammar has more than sqrt(2^31) types -> needs long long for typecache
git-svn-id: https://pet.opendfki.de/repos/pet/main@21 4200e16c-5112-0410-ac55-d7fb557a720a

+++ common/typecache.h
@@ -30,7 +30,7 @@
 
 #include "chunk-alloc.h"
 
-typedef long typecachekey_t;
+typedef long long typecachekey_t;
 
 // note that the initial cache has to fit in one chunk, thus
 // GLB_CACHE_SIZE*sizeof(typecachebucket) must be <= GLB_CHUNK_SIZE
